+++ src/semantic/checks/top_level_declaration.rs
@@ -22,6 +22,7 @@ impl SemanticCheck for Statement<'_> {
         match self {
             Statement::Binding(b) => b.check_semantics(scope),
             Statement::FnDecl(f) => f.check_semantics(scope),
+            Statement::Conditional(_) => unimplemented!("check conditional"),
         }
     }
 }