+++ src/syntax/parsers/mod.rs
@@ -4,3 +4,4 @@ pub mod expression;
 pub mod function_declaration;
 pub mod module;
 pub mod statement;
+pub mod conditional;